+++ lib/presentation/pages/admin_profile/dashboard_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:meta_app/presentation/widgets/header.dart';
+import 'package:meta_app/presentation/widgets/profile_header/header.dart';
 
 class DashboardPage extends StatelessWidget {
   const DashboardPage({super.key});
